Added a y = x plot for the graphs.
The y=x graph, by default, should help with the symlog bits where the linthresh of x and y are different from each other.

The scaling can be a little unusual at times, but given NS Stats with small negative ranges and large positive ranges, symlog is the only way to look at both the small and large values.