Added the Data Extractor.
As part of freedom of data and to help everyone involved, this adds the Data Extractor that I have used and refined eventually in order to get the data I need.

Please note that putting the raw .csv into the Analyser does not work for the time being. I still have some TODOs.